# swagger.yaml
swagger: "2.0"
info:
  description: "This is a simple api documentation."
  version: "1.0.0"
  title: "Swagger ACT"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: ""
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8000"
basePath: "/api"
...
schemes:
- "http"
paths:
...
  /users/{id}: 
    get:
      tags:
      - "User"
      summary: "Get User Details by id"
      description: "Get user details"
      operationId: "UserDetails"
      produces:
      - "application/json"
      parameters: 
        - in: path
          name: id  
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/userById"
    put:
      tags:
      - "User"
      summary: "update User Details by id"
      description: "User update"
      operationId: "UserUpdate"
      produces:
      - "application/json"
      parameters: 
        - in: path
          name: id 
        - in: "body"
          name: "body"
          description: "Update User"
          required: true
          schema:
            properties:
              first_name:
                type: "string"
              last_name:
                type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Success"
    delete:
      tags:
      - "User"
      summary: "Delete User Details by id"
      description: "delte user"
      operationId: "UserDelete"
      produces:
      - "application/json"
      parameters: 
        - in: path
          name: id  
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Success"        
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "Authorization"
    in: "header"
security:
  - api_key: []
...